Refactored Unit Tests to be in separate, more descriptive files
